requirements.txt
# Core Dependencies
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Sources
yfinance==0.2.36
requests==2.31.0

# Technical Analysis
pandas-ta==0.3.14b0

# Visualization
plotly==5.18.0
mplfinance==0.12.10b0
matplotlib==3.8.2

# Dashboard (future)
streamlit==1.31.0

# Development & Testing
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0

# Utilities
tqdm==4.66.1
tabulate==0.9.0
